requirements.txt
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Django==1.11.16
defusedxml==0.7.1
distlib==0.3.6
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.8.0
future==0.16.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.1
oauthlib==3.2.1
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stripe==4.1.0
termcolor==1.1.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.1.9
